Fix AiPersonality GameMode underlying type

+++ src/MphRead.Native/Formats/AiPersonality.hpp
@@ -13,7 +13,7 @@
 
 namespace MphRead
 {
-    enum class GameMode : std::int32_t;
+    enum class GameMode : std::uint8_t;
 }
 
 namespace MphRead::Formats
